+++ kipper/core/src/compiler/parser/file-listener.ts
@@ -55,7 +55,7 @@ import type {
 	TypeSpecifierContext,
 	UnaryOperatorContext,
 } from "./index";
-import { IfStatementContext, SwitchStatementContext } from "./index";
+import { IfStatementContext, VoidOrNullOrUndefinedPrimaryExpressionContext, SwitchStatementContext } from "./index";
 import type { KipperProgramContext } from "../program-ctx";
 import { ParserRuleContext } from "antlr4ts";
 import {
@@ -184,7 +184,7 @@ export class KipperFileListener implements KipperListener {
 	}
 
 	/**
-	 * Handles an incoming statement context. The handling algorithm is:
+	 * Handles an incoming statement context. The handling algorithm is as following:
 	 * - If {@link _currentPrimaryNode} is undefined, then it will be created and set as a child of
 	 * {@link _rootNode}
 	 * - Otherwise, generate a new {@link Statement} instance, which will be added to the {@link _currentPrimaryNode} as
@@ -198,7 +198,7 @@ export class KipperFileListener implements KipperListener {
 	}
 
 	/**
-	 * Handles an incoming statement context. The handling algorithm is:
+	 * Handles an incoming statement context. The handling algorithm is as following:
 	 * - If {@link _currentPrimaryNode} is undefined, then it will be created and set as a child of
 	 * {@link _rootNode}
 	 * - Otherwise, generate a new {@link Declaration} instance, which will be added to the {@link _currentPrimaryNode} as
@@ -212,7 +212,7 @@ export class KipperFileListener implements KipperListener {
 	}
 
 	/**
-	 * Handles an exiting statement or definition context. The handling algorithm is:
+	 * Handles an exiting statement or definition context. The handling algorithm is as following:
 	 * - If {@link _currentPrimaryNode.parent} is of type {@link _currentPrimaryNode} or {@link Statement}, then set
 	 * {@link _currentPrimaryNode} to that parent.
 	 * - Otherwise set {@link _currentPrimaryNode} to {@link undefined} again. If
@@ -433,6 +433,24 @@ export class KipperFileListener implements KipperListener {
 	 * @param ctx The parse tree (instance of {@link ParserRuleContext}).
 	 */
 	public exitBoolPrimaryExpression(ctx: BoolPrimaryExpressionContext): void {
+		this.handleExitingExpressionCtx();
+	}
+
+	/**
+	 * Enter a parse tree produced by the `VoidOrNullOrUndefinedPrimaryExpression`
+	 * labeled alternative in `KipperParser.primaryExpression`.
+	 * @param ctx The parse tree (instance of {@link ParserRuleContext}).
+	 */
+	public enterVoidOrNullOrUndefinedPrimaryExpression(ctx: VoidOrNullOrUndefinedPrimaryExpressionContext): void {
+		this.handleIncomingExpressionCtx(ctx);
+	}
+
+	/**
+	 * Exit a parse tree produced by the `VoidOrNullOrUndefinedPrimaryExpression`
+	 * labeled alternative in `KipperParser.primaryExpression`.
+	 * @param ctx The parse tree (instance of {@link ParserRuleContext}).
+	 */
+	public exitVoidOrNullOrUndefinedPrimaryExpression(ctx: VoidOrNullOrUndefinedPrimaryExpressionContext): void {
 		this.handleExitingExpressionCtx();
 	}