+++ kipper/core/src/compiler/target-presets/target-semantic-analyser.ts
@@ -22,7 +22,7 @@ import type {
 	IdentifierPrimaryExpression,
 	IdentifierTypeSpecifierExpression,
 	IfStatement,
-	IncrementOrDecrementExpression,
+	IncrementOrDecrementPostfixExpression,
 	IncrementOrDecrementUnaryExpression,
 	IterationStatement,
 	JumpStatement,
@@ -41,7 +41,8 @@ import type {
 	VariableDeclaration,
 } from "../semantics";
 import type { CompilableASTNode } from "../parser";
-import { KipperSemanticErrorHandler } from "../semantics";
+import { KipperSemanticErrorHandler, VoidOrNullOrUndefinedPrimaryExpression } from "../semantics";
+import { VoidOrNullOrUndefinedPrimaryExpressionContext } from "../parser";
 
 /**
  * Represents a function that checks the semantics for a {@link CompilableASTNode}.
@@ -158,9 +159,14 @@ export abstract class KipperTargetSemanticAnalyser extends KipperSemanticErrorHa
 	public abstract arraySpecifierExpression: TargetASTNodeSemanticAnalyser<ArraySpecifierExpression>;
 
 	/**
-	 * Performs translation-specific semantic analysis for {@link IncrementOrDecrementExpression} instances.
+	 * Performs translation-specific semantic analysis for {@link VoidOrNullOrUndefinedPrimaryExpression} instances.
 	 */
-	public abstract incrementOrDecrementExpression: TargetASTNodeSemanticAnalyser<IncrementOrDecrementExpression>;
+	public abstract voidOrNullOrUndefinedPrimaryExpression: TargetASTNodeSemanticAnalyser<VoidOrNullOrUndefinedPrimaryExpression>;
+
+	/**
+	 * Performs translation-specific semantic analysis for {@link incrementOrDecrementPostfixExpression} instances.
+	 */
+	public abstract incrementOrDecrementPostfixExpression: TargetASTNodeSemanticAnalyser<IncrementOrDecrementPostfixExpression>;
 
 	/**
 	 * Performs translation-specific semantic analysis for {@link FunctionCallPostfixExpression} instances.