+++ kipper/target-js/src/semantic-analyser.ts
@@ -20,7 +20,7 @@ import type {
 	GenericTypeSpecifierExpression,
 	IdentifierPrimaryExpression,
 	IdentifierTypeSpecifierExpression,
-	IncrementOrDecrementExpression,
+	IncrementOrDecrementPostfixExpression,
 	IncrementOrDecrementUnaryExpression,
 	IterationStatement,
 	JumpStatement,
@@ -38,7 +38,12 @@ import type {
 	TypeofTypeSpecifierExpression,
 	VariableDeclaration,
 } from "@kipper/core";
-import { IfStatement, KipperTargetSemanticAnalyser, ReservedIdentifierOverwriteError } from "@kipper/core";
+import {
+	IfStatement,
+	KipperTargetSemanticAnalyser,
+	ReservedIdentifierOverwriteError,
+	VoidOrNullOrUndefinedPrimaryExpression,
+} from "@kipper/core";
 import { getJavaScriptBuiltInIdentifier } from "./tools";
 
 /**
@@ -251,9 +256,14 @@ export class JavaScriptTargetSemanticAnalyser extends KipperTargetSemanticAnalys
 	arraySpecifierExpression = async (node: ArraySpecifierExpression) => {};
 
 	/**
-	 * Performs typescript-specific semantic analysis for {@link IncrementOrDecrementExpression} instances.
+	 * Performs typescript-specific semantic analysis for {@link VoidOrNullOrUndefinedPrimaryExpression} instances.
 	 */
-	incrementOrDecrementExpression = async (node: IncrementOrDecrementExpression) => {};
+	voidOrNullOrUndefinedPrimaryExpression = async (node: VoidOrNullOrUndefinedPrimaryExpression) => {};
+
+	/**
+	 * Performs typescript-specific semantic analysis for {@link IncrementOrDecrementPostfixExpression} instances.
+	 */
+	incrementOrDecrementPostfixExpression = async (node: IncrementOrDecrementPostfixExpression) => {};
 
 	/**
 	 * Performs typescript-specific semantic analysis for {@link FunctionCallPostfixExpression} instances.